Code fixes for the Rubocop warnings.
These are the _easy_ fixes to the code structure, and many were "auto-corrected"
by Rubocop.

The real changes (long and complex methods) will need more analysis of the code.

Also fixed the `rdoc` generation to use the correct title and main file.

Also added an auto-generated rubocop_todo config to defer fixes.

This setup lets the issues be documented, while not becoming a nag.