Removed the `is_` and `has_` prefixes from predicate methods.
This is to comply with the Ruby method naming convention for predicate methods
(the ones ending with `?`). Rest of the code base has been updated to use the
new method names.

However, for backward compatibility, `aliases` with the old method names have
been defined to the new names so that current client code can continue to work.